// src/app/model/ISubleaseTovTermiante.ts
export default interface ISubleaseTovTerminate{
    NumberGroup: number;
    NameGroup: string;
    ContractNumber: string;
    CreationContractDate: Date;
    CreationDate: Date;
    PipSublessor: string;
    rnokppSublessor: string;
    addressSublessor:string;
    PipDirector:string;
    PipsDirector:string
    EndContractData: Date;
    RoomArea:number;
    RoomAreaText: string;
    RoomAreaAddress:string;
    BanckAccount: string;
}